Update minimap-titles.coffee
added feature to toggle comment borders on or off, changed coffeescript from a block quote to single line comments followed by a space (this can just have an if statement if we want the old block comment of ### functionality), added dynamic sizing of borders to be based on atom editors preferred width based on the config.cson setting and to use the default if it isn't supplied (currently 80). tested all cases of code to ensure appropriate output, updated menu to add the toggle comment borders command.